feat: item info repository 추상화를 위한 Port, PortImpl 클래스 분리
