+++ src/main/java/until/the/eternity/iteminfo/domain/repository/ItemInfoRepositoryPort.java
@@ -1,12 +1,10 @@
 package until.the.eternity.iteminfo.domain.repository;
 
 import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
 import until.the.eternity.iteminfo.domain.entity.ItemInfo;
 
-public interface ItemInfoRepository extends JpaRepository<ItemInfo, String> {
-
-    List<ItemInfo> findByNameContaining(String name);
+public interface ItemInfoRepositoryPort {
+    List<ItemInfo> findAll();
 
     List<ItemInfo> findByTopCategory(String topCategory);
